+++ R/replace.R
@@ -1,8 +1,15 @@
 #' @useDynLib covr covr_duplicate_
-replacement <- function(name, env = as.environment(-1)) {
-  target_value <- get(name, envir = env)
+replacement <- function(name, env = as.environment(-1), target_value = get(name, envir = env)) {
   if (is.function(target_value) && !is.primitive(target_value)) {
-    new_value <- trace_calls(target_value, name)
+    if (is_vectorized(target_value)) {
+      new_value <- target_value
+      environment(new_value)$FUN <- trace_calls(environment(new_value)$FUN, name)
+    } else if (is.function(target_value) && inherits(target_value, "memoised")) {
+      new_value <- target_value
+      environment(new_value)$`_f` <- trace_calls(environment(new_value)$`_f`, name)
+    } else {
+      new_value <- trace_calls(target_value, name)
+    }
     attributes(new_value) <- attributes(target_value)
 
     if (isS4(target_value)) {