+++ NEWS.md
@@ -1,3 +1,13 @@
+## 2.1.0 ##
+* Add support for GitLab CI (#190, @enbrown).
+* Update exclusion documentation to include line_exclusions and function
+  exclusions (#191).
+* Support coverage of R6 methods (#174).
+* Explicitly set default packages (including methods) (#183, #180)
+* Set R_LIBS and R_LIBS_SITE as well as R_LIBS_USER (#188).
+* Automatically exclude RcppExport files (#170).
+* Memoised and Vectorized functions now able to be tracked.
+
 ## 2.0.1 ##
 * Support for filtering by function as well as line.
 * Now tracks coverage for RC methods